+++ readium/streamer/src/main/java/org/readium/r2/streamer/parser/epub/ManifestAdapter.kt
@@ -58,6 +58,16 @@ internal class ManifestAdapter(
                 }
             }
             .mapValues { listOf(PublicationCollection(links = it.value)) }
+            .toMutableMap()
+
+        // EPUB 3 Reading Systems must ignore the guide element when provided in EPUB 3 Publications
+        // whose EPUB Navigation Document includes the landmarks feature.
+        // https://idpf.org/epub/30/spec/epub30-publications.html#sec-guide-elem
+        if (!subcollections.contains("landmarks") && packageDocument.guide.isNotEmpty()) {
+            // EPUB 2.0 doesn't have a landmarks collection, so we use the guide as a fallback
+            // If an EPUB 3.0+ file does not have landmarks, it will use guide instead.
+            subcollections["landmarks"] = listOf(PublicationCollection(links = packageDocument.guide))
+        }
 
         // Build Publication object
         return Manifest(
@@ -66,7 +76,7 @@ internal class ManifestAdapter(
             readingOrder = readingOrder,
             resources = resources,
             tableOfContents = toc,
-            subcollections = subcollections
+            subcollections = subcollections.toMap()
         )
     }
 }